chatapp-backend/requirements.txt
django
django-cors-headers
django-rest-framework
whitenoise
channels
daphne
pymongo
python-dotenv
djongo
certifi